refactor(trade-ai): extract shared validation into _shared_validation package
Both trade_insights_ai and trade_blast had near-identical copies of 10
validator/leniency files (~2,300 lines) differing only in import paths.
Extract into reports/_shared_validation/ so both lanes import from one
source. Eliminates maintenance hazard where a fix in one lane wouldn't
propagate to the other.

- New _shared_validation/ with constants, util, leniency/*, validator_rules/*
- Both lane packages now import shared code; lane-specific files stay put
- prompt_text.py in both lanes imports constants from shared instead of
  defining them inline
- 925 unit + 362 web tests pass, no behavioral change
