+++ src/uw_scan/reports/_shared_validation/leniency/__init__.py
@@ -1,0 +1,1 @@
+"""Shared leniency coercers for trade_insights_ai and trade_blast."""
