+++ src/uw_scan/reports/trade_blast/leniency/__init__.py
@@ -1,9 +1,9 @@
 """Claude leniency helpers for Trade Insights AI."""
 
+from uw_scan.reports._shared_validation.leniency.triggers import _coerce_option_legs
 from uw_scan.reports.trade_blast.leniency.coerce import (
     _coerce_claude_outcome_dict,
 )
 from uw_scan.reports.trade_blast.leniency.framework import _coerce_framework
-from uw_scan.reports.trade_blast.leniency.triggers import _coerce_option_legs
 
 __all__ = ["_coerce_claude_outcome_dict", "_coerce_framework", "_coerce_option_legs"]
