+++ src/uw_scan/reports/trade_insights_ai/leniency/__init__.py
@@ -1,8 +1,8 @@
 """Claude leniency helpers for Trade Insights AI."""
 
+from uw_scan.reports._shared_validation.leniency.triggers import _coerce_option_legs
 from uw_scan.reports.trade_insights_ai.leniency.coerce import (
     _coerce_claude_outcome_dict,
 )
-from uw_scan.reports.trade_insights_ai.leniency.triggers import _coerce_option_legs
 
 __all__ = ["_coerce_claude_outcome_dict", "_coerce_option_legs"]
