--- src/uw_scan/reports/trade_insights_ai/validator_rules/imperative.py
@@ -1,47 +1,0 @@
-"""Imperative-text safety checks for Trade Insights AI outcomes."""
-
-from __future__ import annotations
-
-import json
-import re
-
-from uw_scan.models import TradeInsightAiOutcome
-
-_IMPERATIVE_PHRASES = (
-    "buy now",
-    "sell now",
-    "enter now",
-    "execute this trade",
-    "place this order",
-    "size this position",
-)
-_IMPERATIVE_PATTERNS = tuple(
-    re.compile(pattern, re.I)
-    for pattern in (
-        r"\byou\s+should\s+(?:buy|sell|enter|open|execute)\b",
-        r"\bmust\s+(?:buy|sell|enter|open|execute)\b",
-        r"\b(?:take|open)\s+this\s+trade\b",
-        r"\b(?:place|send)\s+(?:this|the|an?)\s+order\b",
-        r"\bgo\s+(?:long|short)\s+now\b",
-    )
-)
-
-def _reject_imperative_text(outcome: TradeInsightAiOutcome) -> None:
-    checked = [
-        outcome.headline.stance_label,
-        outcome.preferred_expression.title if outcome.preferred_expression else "",
-        outcome.preferred_expression.subtitle if outcome.preferred_expression else "",
-    ]
-    for text in checked:
-        lowered = text.strip().lower()
-        if any(lowered.startswith(phrase) for phrase in _IMPERATIVE_PHRASES):
-            raise ValueError(f"imperative trade instruction rejected: {text}")
-
-    free_text = json.dumps(outcome.model_dump(mode="json"), default=str)
-    for phrase in _IMPERATIVE_PHRASES:
-        if re.search(rf"(^|[.!?]\s+){re.escape(phrase)}\b", free_text, re.I):
-            raise ValueError(f"imperative trade instruction rejected: {phrase}")
-    for pattern in _IMPERATIVE_PATTERNS:
-        if pattern.search(free_text):
-            raise ValueError("imperative trade instruction rejected")
-
